backend/requirements.txt
# FastAPI + server
fastapi==0.115.0
python-multipart>=0.0.9
# Error monitoring (optional — set SENTRY_DSN to activate)
sentry-sdk[fastapi]>=2.0
resend>=2.0
uvicorn[standard]==0.32.0
sse-starlette==2.1.0
python-jose[cryptography]==3.3.0
email-validator==2.2.0
langgraph-checkpoint-sqlite==3.0.3
openpyxl>=3.1
# Redis (optional — set REDIS_URL to activate multi-pod run state + rate limiting)
redis>=5.0
# Optional — set DATABASE_URL to activate:
# langgraph-checkpoint-postgres

# Project deps (minus streamlit)
altair==6.0.0
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
duckdb==1.5.0
fpdf2==2.8.7
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
holidays==0.93
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
kiwisolver==1.5.0
langchain-core==1.2.20
langfuse==4.0.0
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.20
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
narwhals==2.18.0
networkx==3.6.1
numpy==2.4.3
openai==2.29.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-sdk==1.40.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.3
pillow==12.1.1
pluggy==1.6.0
prophet==1.3.0
protobuf==6.33.6
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==82.0.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
wrapt==1.17.3
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
